fix: measure hot-path loudness as real K-weighted LUFS, not RMS - 23 dB (#5221)
`calculate_loudness_units()` returned `to_db(rms(audio)) - 23.0`. Both halves
were wrong. There is no K-weighting in plain RMS, and -23.0 is the EBU R128
*target* level (`EBU_R128_TARGET_LUFS` in assessment_constants.py), not a
dBFS->LUFS conversion — BS.1770-4's absolute-scale constant is -0.691, which
`loudness_meter.calculate_block_loudness()` and the fingerprint's own proxy in
`fingerprint/windowed_compute.py` both already use. Measured against the
codebase's own BS.1770-4 meter, the function read 24-29 LU low.

This was not merely imprecise, because every consumer is calibrated in real
LUFS and only this function spoke the broken scale:

  - `continuous_dsp_ops._apply_final_normalization` computes
    `adjustment = target_lufs - current_lufs`, where `target_lufs` derives
    from the fingerprint's -0.691-scale `lufs` and from
    `PresetProfile.target_lufs` (-14.0 etc.). On a representative musical
    buffer the two sides sat 25.5 dB apart, so the step applied ~+25 dB of
    spurious gain that Step 2 then clawed back to the -0.3 dBFS ceiling —
    "LUFS normalization" was behaving as peak normalization, the very thing
    #3665 was filed to stop.
  - `AdaptiveLoudnessControl` compares against `TARGET_LUFS = -11.0` and
    `VERY_LOUD_THRESHOLD = -12.0`; readings near -40 made the "already loud"
    branch unreachable.
  - `adaptive_mode`'s `loudness_coordinate` tanh is centred on -14.3887; it
    sat at ~0.0005 instead of ~0.45, so the compression/expansion blend it
    drives was effectively dead.

So this is a calibration repair, not a rescaling: the downstream constants
were always written for true LUFS and need no re-tuning, which resolves the
issue's RETURN VALUE completeness check. All call sites move together because
they share this one function (CONSISTENCY check).

The function now delegates to `LoudnessMeter`'s BS.1770-4 K-weighting rather
than duplicating filter design, and matches the meter to <=0.1 LU. A fresh
meter per call: `apply_k_weighting()` carries filter state between calls, so a
shared instance would let one buffer colour the next, and the pipeline
measures unrelated buffers from several threads. Gating is deliberately not
applied — it is undefined for the sub-400 ms buffers this is called with, and
moves a full chunk by <0.05 LU at 4.6x the cost; the docstring says so and
points at `LoudnessMeter` for gated integrated loudness.

DTYPE: activating the makeup-gain branch surfaced a latent float32 -> float64
upcast in `PeakNormalizer.normalize_to_target`, which scaled by a np.float64
scalar (`calculate_adaptive_peak_target` returns one). float()-ing the scalar
restores the dtype invariant for every caller.

Cost: ~45 ms per 15 s stereo chunk vs ~0.5 ms, ~0.3% of real time per call.

Tests: 903 passed across tests/auralis/{core,dsp,analysis} (the one failure,
test_similarity_fit, is pre-existing and listed in pytest-baseline.json), plus
25 passed in tests/test_preset_system.py and 10 passed in the end-to-end
mastering regression suite. New: 8 tests pinning meter agreement, K-weighting
(equal-RMS tones at 50/1k/8k must not read equally loud), round-tripping a
preset target, the silence floor, and no filter-state bleed between calls.

+++ auralis/core/processing/base/peak_management.py
@@ -91,7 +91,13 @@ class PeakNormalizer:
 
         if peak > 0.001:  # Avoid division by zero
             target_peak = DBConversion.to_linear(target_peak_db)
-            audio = audio * (target_peak / peak)
+            # float() the scalar before scaling. NumPy promotes
+            # float32_array * np.float64_scalar to float64, and callers do
+            # hand np.float64 targets through here — e.g.
+            # AdaptiveLoudnessControl.calculate_adaptive_peak_target() returns
+            # one — which silently broke the float32-in/float32-out dtype
+            # invariant. A Python float keeps the array's own dtype.
+            audio = audio * float(target_peak / peak)
             ProcessingLogger.post_stage("Peak Normalization", peak_db, target_peak_db, "Peak")
             return audio, peak_db
